% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Adjacency matrix}
\usage{
adjacency_matrix(object)
}
\arguments{
\item{object}{igraph object nebo maticový zápis OTU tabulky}
}
\value{
maticový zápis adjacenční matice
}
\description{
Funkce adjacency_matrix převede igraph objekt nebo OTU tabulku do adjacenční matice.
}
