% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph.R
\name{get_graph}
\alias{get_graph}
\title{Get graph}
\usage{
get_graph(object, weighted = TRUE, community = "walktrap",
  bipartite_projection = FALSE)
}
\arguments{
\item{object}{maticový zápis OTU tabulky nebo adjacenční matice}

\item{weighted}{logická proměnná. Přednastavená volba weighted = TRUE iniciuje váhování hran. Nastavení weighted = FALSE odstraní váhování hran.}

\item{community}{textová proměnná. Metoda detekce komunit. Možné volby jsou: walktrap, spinglass, leading_eigen, label_prop,
fast_greedy, edge_betweennes, multilevel, infomap, optimal}

\item{bipartite_projection}{logická proměnná. Umožňuje využití projekce bipartitního grafu pro detekci komunit. Volby bipartite_projection = 1 a bipartite_projection = 2
vedou k využití jedné ze dvou bipartitních projekcí, volby bipartite_projection = FALSE zakáže tvorbu bipartitních projekcí.}
}
\value{
objekt igraph
}
\description{
Funkce get_graph vytváří z OTU tabulek nebo adjacenčních matic igraph objekty, které jsou také rovnou vykresleny.
}
