% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(biom_file)
}
\arguments{
\item{biom_file}{již načtený biom soubor}
}
\value{
dostupná metadata a počet levlů v nich
}
\description{
Funkce get_metadata vypisuje dostupná metadata a počet levlů v nich.
}
