% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OTUtable.R
\name{get_OTUtable}
\alias{get_OTUtable}
\title{Get OTUtable}
\usage{
get_OTUtable(biom_file, taxonomy = "class", NewOTU = FALSE,
  metadata = NULL, threshold = 10)
}
\arguments{
\item{biom_file}{již načtený biom soubor}

\item{taxonomy}{textová proměnná. Taxonomická kategorie, na základě, které mají být data redukována. Možné volby jsou:
kingdom, phylum, class, order, family, genus a species.}

\item{NewOTU}{logická proměnná. Nastavení NewOTU = TRUE iniciuje zachování všech OTU včetně těch, u nichž není známá požadovaná taxonomická kategorie.
Přednastavená volba NewOTU = FALSE vede k odstranění OTU bez požadované taxonomické kategorie z dalšího zpracování.}

\item{metadata}{textová proměnná. Název metadat na základě kterých mají být data porovnávána a redukována}

\item{threshold}{číselná proměnná. Prahová hodnota se kterou jsou pozorování porovnávána.}
}
\value{
maticový zápis redukované OTU tabulky
}
\description{
Funkce get_OTUtable využívá tři přístupy k redukci OTU tabulky: Redukce podle taxonomie, redukce podle metadat (pokud jsou k dispozici)
a redukce nastavením prahové hodnoty.
}
