% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metadata.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge metadata}
\usage{
merge_metadata(biom_file, OTUtab, metadata)
}
\arguments{
\item{biom_file}{již načtený biom soubor}

\item{OTUtab}{maticový zápis OTU tabulky}

\item{metadata}{textová proměnná. Název metadat na základě kterých mají být data porovnávána a redukována}
}
\value{
maticový zápis redukované OTU tabulky
}
\description{
Funkce merge_metadata redukuje OTU tabuku na základě dostupných metadat. OTU shodující se v metadatech jsou
sloučena do jednoho OTU nazvaného po hodnotě metadat.
}
