% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxonomy.R
\name{merge_taxonomy}
\alias{merge_taxonomy}
\title{Merge taxonomy}
\usage{
merge_taxonomy(biom_file, OTUtab, taxonomy, NewOTU = FALSE)
}
\arguments{
\item{biom_file}{již načtený biom soubor}

\item{OTUtab}{maticový zápis OTU tabulky}

\item{taxonomy}{textová proměnná. Taxonomická kategorie, na základě, které mají být data redukována. Možné volby jsou:
kingdom, phylum, class, order, family, genus a species.}

\item{NewOTU}{logická proměnná. Nastavení NewOTU = TRUE iniciuje zachování všech OTU včetně těch, u nichž není známá požadovaná taxonomická kategorie.
Přednastavená volba NewOTU = FALSE vede k odstranění OTU bez požadované taxonomické kategorie z dalšího zpracování.}
}
\value{
maticový zápis redukované OTU tabulky
}
\description{
Na základě požadované taxonomické úrovně jsou porovnávány OTU mezi sebou
a OTU se stejnou taxonomickou kategorií jsou sečtena a přiřazena novému OTU nazvanému podle taxonomického určení,
které je spojuje
}
