% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_adjust.R
\name{name_adjust}
\alias{name_adjust}
\title{Name adjust}
\usage{
name_adjust(object)
}
\arguments{
\item{object}{igraph object, maticový zápis OTU tabulky nebo adjacenční matice}
}
\value{
igraph object, maticový zápis OTU tabulky nebo adjacenční matice se zkrácenými názvy
}
\description{
Funkce name_adjust slouží ke zkrácení automaticky QIIME vygenerovaného názvu taxonomických kategorií a OTU.
}
