% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_gml.R
\name{save_as_gml}
\alias{save_as_gml}
\title{Save as gml}
\usage{
save_as_gml(object, name = "graf.gml")
}
\arguments{
\item{object}{igraph object, maticový zápis OTU tabulky nebo adjacenční matice}

\item{name}{textová proměnná. Název nového gml souboru.}
}
\description{
Funkce save_as_gml umožňuje převedení OTU tabulek, adjacenčních matic nebo igraph objektů do bipartitních grafů v gml formátu.
}
