% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_focus.R
\name{taxonomy_focus}
\alias{taxonomy_focus}
\title{Taxonomy focus}
\usage{
taxonomy_focus(biom_file, OTUtab, taxlevel, value)
}
\arguments{
\item{biom_file}{již načtený biom soubor}

\item{OTUtab}{maticový zápis OTU tabulky}

\item{taxlevel}{textová proměnná. Taxonomická kategorie, na základě, které mají být data redukována. Možné volby jsou:
kingdom, phylum, class, order, family, genus a species.}

\item{value}{textová proměnná. Specifická hodnota taxonomie jako požadovaná vlastnost všech OTU ve výstupní OTU tabulce.}
}
\value{
maticový zápis redukované OTU tabulky
}
\description{
Funkce taxonomy_focus umožňuje zaměřit se pouze na určitou taxonomickou kategorii. Všechny OTU nenáležející
do této specifické kategorie jsou z OTU tabulky odstraněny.
}
